Added support for Option types in runtime_object.go
The "Option" type was introduced to runtime_object.go along with functions to create a new Option type and to unwrap it. Also, boolean functions to check various types of the runtime value were added, and "Null" was renamed to "None". This would help in better handling expressions and precedence.