Update lexer token matching with performance improvements
Improved the lexer token matching logic to be more efficient and easier to maintain. Implemented the readNext() function which uses the new matchSymbols() and readIdentifier() functions. This makes the lexer able to match symbols and identifiers without consuming any input. Changes in this area, along with updates to the advance() method, should provide performance improvements. Also, added support for the uppercase version of keywords and ordered the keyword matching table based on the length of the value for quicker access.